+++ src/ecutils/utils/__init__.py
@@ -1,0 +1,5 @@
+# ecutils/utils/__init__.py
+
+from ecutils.utils.settings import LRU_CACHE_MAXSIZE
+
+__all__ = ["LRU_CACHE_MAXSIZE"]